LayoutTests/js/regexp-divequal-expected.txt

Test JS parser handling of regex literals starting with /=

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /=/.toString() is '/=/'
PASS /=/.test('') is false
PASS /=/.test('=') is true
PASS '='.match(/=/) is ['=']
PASS '='.match(/\=/) is ['=']
PASS successfullyParsed is true

TEST COMPLETE
